requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
flake8==6.0.0
Markdown==3.4.1
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pytz==2022.7.1
sqlparse==0.4.3
